backend/requirements.txt
flask
flask-cors
opencv-python
anthropic
python-dotenv
pytesseract
Pillow
numpy
